docs: Improved structure for GP types
Also added documentation for some of them
