+++ .prettierignore
@@ -2,3 +2,4 @@ package-lock.json
 pnpm-lock.yaml
 yarn.lock
 *.md
+docs/**
