+++ eslint.config.mjs
@@ -76,5 +76,8 @@ export default [
     rules: {
       "import/no-anonymous-default-export": "off"
     }
+  },
+  {
+    ignores: ["docs/**"]
   }
 ];
