+++ src/hooks.ts
@@ -56,7 +56,7 @@ export const WindowContext = createContext<Window>(window);
  * or `undefined` if no message has been sent yet.
  *
  * Component rerenders on receiving a new message.
- * 
+ *
  * @category React hooks
  *
  * @param typeFilter The type of messages to listen for.
@@ -120,7 +120,7 @@ function sendMessage(message: SendMessage) {
  * that has been sent, or `null` if no such message has been sent.
  *
  * Component rerenders on receiving a new message.
- * 
+ *
  * @category React hooks
  */
 export function useGraphData() {
@@ -132,7 +132,7 @@ export function useGraphData() {
  * that has been sent, or `null` if no such message has been sent.
  *
  * Component rerenders on receiving a new message.
- * 
+ *
  * @category React hooks
  */
 export function useMLData() {
@@ -141,10 +141,10 @@ export function useMLData() {
 
 /**
  * A react hook that fetches the configuration for this addon's instance from the GraphPolaris session. Component rerenders on receiving a new configuration.
- * 
+ *
  * @remarks
  * This hook should only be used for the visualization component. For setting components use the {@link useSettings} hook that can also push a new configuration to GraphPolaris.
- * 
+ *
  * @example
  * ```tsx
  * type VisSettings = { theme: "dark" | "light" };
@@ -156,11 +156,11 @@ export function useMLData() {
  *
  * @template T
  * The type of the configuration which must adhere to the configuration requirements.
- * 
+ *
  * @returns
  * Returns the last message containing visualization settings
  * that has been sent, or `null` if no configuration has yet been sent.
- * 
+ *
  * @category React hooks
  * @category Settings
  */
@@ -173,7 +173,7 @@ export function useSettingsData<T extends Settings>(): T | undefined {
  *
  * @remarks
  * This hook should only be used for the settings component. The update function's messages are ignored by GraphPorlaris if it is called from the visualisation component, use the {@link useSettingsData} hook instead.
- * 
+ *
  * @example
  * ```tsx
  * type VisSettings = { theme: "dark" | "light" };
@@ -189,10 +189,10 @@ export function useSettingsData<T extends Settings>(): T | undefined {
  *   </>);
  * }
  * ```
- * 
+ *
  * @category React hooks
  * @category Settings
- * 
+ *
  * @template T
  * The type of the configuration which must adhere to the configuration requirements.
  *
@@ -229,7 +229,7 @@ export type UpdateFunction<T> = (changes: Partial<T>) => void;
  * that has been sent, or `null` if no such message has been sent.
  *
  * Component rerenders on receiving a new message.
- * 
+ *
  * @category React hooks
  *
  * @returns A `SchemaGraph` containing the schema of the currently selected data.
