+++ src/types/index.ts
@@ -1,0 +1,25 @@
+/*
+ * This program has been developed by students from the bachelor
+ * Computer Science at Utrecht University within the Software Project course.
+ *
+ * © Copyright Utrecht University
+ * (Department of Information and Computing Sciences)
+ */
+
+/*
+ * This program has been developed by students from the bachelor
+ * Computer Science at Utrecht University within the Software Project course.
+ *
+ * © Copyright Utrecht University
+ * (Department of Information and Computing Sciences)
+ */
+
+/**
+ * @packageDocumentation
+ *
+ * @categoryDescription Graph data
+ * These types are used for transferring raw graph data.
+ */
+
+export * from "./graphpolaris.types";
+export * from "./graphpolaris.graphQueryResult.types";
